remove all fallback code
I get reports of people missing the reference part of the text. This
is likely caused by a fallback path. The fallback path is poorly tested
and hard to maintain. At the same time, AELF API and website greatly
improves their quality. It is time for this old code to retire.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>