Fix completely broken start and start instruction
During the 2-3 years when I did not maintain the application, the code
suffered dramatic bit-rot. It was no longer possible to build the Docker
image are wrong and the tests are completely broken since nose test is
abandonned.

This commit bumps all dependencies of the application and the Docker
image. Surprisingly, it looks like it still works. The tests were not so
lucky, and required migrating away from the un-maintained "nose"
framework.

Regarding the tests, a lot of warnings remain that must be tackled BUT,
at least, we are back on track.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>

Migrate tests to pytest

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>