Migrate from docker swarm to docker context
Docker Swarm is complex and we are not really using it anyway. In
practice, we just build and run container images on each nodes without
coordination nor private networks, nor HA, ...

Using Docker Swarm requires a coordinator node, which may be down and
coordination when bringing up a new node.

On the opposite, docker context is essentially a replacement from Docker
Machine and offers the ability to control remote Docker daemons
transparently. This is just enough for our use case.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>