Refactor deploy script to support retries and target filter
The retries are needed to help improve deployment consistency. Before
this change, one-off errors on one of the machine could cause the deploy
to only reach a subset of the machines. This is still not perfect as the
script still reports a success in case of errors, but issue likelihood
is now reduced.

Second issue was that the script was only working on all machines, with
not possible restriction. This is a problem when testing a configuration
in a specific machine before applying to all 3 machines. This commit
adds a '--limit' flag inspired by Ansible for this purpose.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>