Add memory limit to the containers
The application is leaking memory, which is bad. What is even worse is
that we let it leak until the OOM killer has to kick-in which renders
the server unavailable for much longer than if we just let it die
prematurely and restart. Of course, a future step will need to rework
allocations.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>