Add new cache-friendly 'checksums' route
On every sync, the phones fetch all offices for the next 7 days, even if
a copy is already present in the local cache. This is done to catch any
change or fix in the office text. Indeed, corrections are typically done
a few days in advance, or even the day itself.

While this approach guarantees fresh data, it also consumes far more BW
than necessary and also un-necessarilly stresses the servers.

Instead, we'll move to a model where the phones first query the list of
checksums to expect over the next 7 days (or less). This request and
response can be efficiently cached by Nginx. Then, the phones will
compare the checksums with the checksums in the local cache and skip the
other requests if the cache is up to date.

In a typical scenario, where nothing changed, this is expected to
replace 63 out of the 72 daily requests per phone, by a single one.
That is, 86% less upstream requests!

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>