Make Etag logic follow the standard
RFC 7232 section 2.3 mandates doubles quotes around the Etag value. This
was not the case up to now. The root cause for this mistake was that the
code was manually setting the header rather than making use of the
built-in helpers.

This fix is not expected to have negative side effects as the feature is
still not used by applications in the wild and was noticed precisely
during the implementation.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>