Add generation-date field to the cache manifest
The cache manifest API and individual offices are eventually consistent.
This is especially true with the Nginx cache that is valid for 4 hours.
In the worse cache, an office may be rereshed just after the cache
manifest in generated. If a phone gets this value, and immediately
refreshes, it will believe it has an outdate version and re-fetch... the
already up to date most recent version it already has. And would do so
for 4hs (if it was in a loop).

This is even more important with the 3 workers.

To break such ties without sacrificing the cache, this commit extends
the API to report the cache generation date in addition to the checksum
itself.

Another approach to avoid changing the API could have been to include
the timestamp in the MSB of the checksum so that the checksums would
become sortable in lexicographioc order. While "clever" this was not
done this way as introducing an object in place of a field also makes
the API extensible in the future.

Moreover, this was the occasion to to turn this API into a more generic
and future proof meta-data API.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>