Add copyright mention as reference when provided
AELF API provides copyright notice for third party texts like the hymns.
For some reason these were never displayed in the application. We are
lucky enough that nobody ever complained...

This commit forwards the raw values and merges them in the 'reference'
field so that applications in the field will get them, even without
upgrading.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>