Support antienne propagation for psalms split into sections
In the general case, each psalm of an office gets its own antienne.
However, in somes cases, the antienne applies to multiple psalm
sections. When this is the case, the antienne should be sung at the
beginning of the first psalm and the end of the last one.

The server side now detects this case and propagates a copy of the
antienne over the various sections. This will allow displaying the
antienne in any section in already deployed version.

Additionally, for a future version of the application, this commit
injects an additional meta-data to inform the application of the
position in the list of sections: first, last or intermediate.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>