Refactor office antienne API
Using feedback from testers and backend implementation, this commit
refactors the API. The main driver for this change is that we'll
introduce another dedicated field for controlling doxology visibility.
It turns out that is does follow exactly the same logic. The renaming of
the field will allow a more consistent API design. It also officializes
the move of more logic / domain knowlege to the backend where we can
more easily iterate.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>