Support correct logic to decide when to place the Doxology
As a general rule, a doxology shall be inserted after psalms and cantic
with the following exceptions:

* Dn 3 which implies the Doxology
* End of a non final psalm section, when a psalm is split AND there is
  no dedicated antienne.

For instance, when we have the '118-18, 87-I, 87-2' sequence in the
Sexte office with a single shared antienne, the Doxology shall be
inserted after te first and the last psalm but not the intermediate.
However, if there are dedicated antiennes, the Doxology is inserted in
every place.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>