Overhaul the 'informations' section
The 'informations' section of the data returned by the API was barely
used until now. An upcoming feature of the application will use the
information embedded in individual offices to generate a drawer header.
This requires data with a structure cleaner that the raw data currently
provided by AELF.

This commit implements the extraction and cleanup logic. Since this is
very brittle, a full year worth of the API has been dumped and is being
used as the basis of unit tests to ensure good enough quality.

Most legacy fields are cleaned in the process. The 'text' field is
preserved for now even though it is completely useless. In the future,
we'll use it to encourage users to upgrade to get cleaner information.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>