Shorten the day description in the general case
In the general case, the name is of the form 'day_name, Nth week of the
ordinary time' where the day name is redundant with the lirturgical name
and increases the likelyhood of a line wrap on the phone.

When in this case, let's just display the week, which is the information
that is not already presented.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>