Kill the RSS API
A first attempt to kill the RSS API was made a few month ago and did not
go well. At least 1k users still rely on it from a very old version.

During the past few month, we prepared a new version that would bring
back reliable offline features as well as support for older phones and
the deprecated LetsEncrypt root CA.

Now that all this work is in place, we can kill this API. Of course,
this is when I realized that all tests still relied on it. Great.

Note: We do not provide an upgrade link because the minimum version to
support links is 2.9.0, unfortunately.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>