feat: terrible hack to display the normal mass on the last sunday of October
Each Church has its own dedication celebration. In rare circumstances,
the date is unknown. For these rare case, the default date is the last
Sunday of October.

Since this is an exception and not the nominal case, I get tons of mails
every year notifying me of a "mistake". This year, let's try something
different and prepend a "variant" with the mass from the "romain"
region.

From a user point of view, this will show as a "variant", just like on
easter. From a code point of view, this is a terrible hack that involves
a circular dependency. I'm ashamed and whish this was not Open Source
and nobody would ever see it. Unfortunately, I have too littel spare
time.

Please forgive me.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>